infra: logic encapsulation, refactoring
