--- api/index.ts
@@ -1,5 +1,0 @@
-import environment from "@/api/environment";
-
-export default {
-  environment,
-};
