+++ pages/Dashboard/api/index.ts
@@ -1,0 +1,5 @@
+import environment from "@/pages/Dashboard/api/environment";
+
+export default {
+  environment,
+};
